# sur-frontend/src/lib/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "EmotionLabel": {
        "enum": [
          "anger",
          "sadness",
          "happiness",
          "fear",
          "surprise",
          "neutral"
        ],
        "title": "EmotionLabel",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
      "RegenerateRequest": {
        "description": "POST /api/segments/{id}/regenerate.\n\n`stages` lets the caller ask for exactly what changed: an edited\ntranslation or an emotion override needs [\"synthesize\"] (re-translating\nwould throw the edit away); [\"translate\", \"synthesize\"] re-derives the\ntext too. The export is re-muxed afterwards either way.\n\nValidated, not filtered: unknown stage names used to be dropped silently\n(and an empty result fell back to the default), and [\"translate\"] alone\nleft the rendered audio saying the old line.",
        "properties": {
          "stages": {
            "items": {
              "enum": [
                "translate",
                "synthesize"
              ],
              "type": "string"
            },
            "minItems": 1,
            "title": "Stages",
            "type": "array"
          }
        },
        "title": "RegenerateRequest",
        "type": "object"
      },
      "SegmentPatch": {
        "description": "PATCH /api/segments/{id} -- edit translated text or override emotion.\n\nAll fields optional; only provided fields are updated. Editing either\nfield does NOT itself trigger a re-render -- call /regenerate for that,\nso a reviewer can queue up several text edits before spending a TTS call.",
        "properties": {
          "emotion_label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmotionLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "translated_text": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translated Text"
          }
        },
        "title": "SegmentPatch",
        "type": "object"
      },
      "SegmentRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detected_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Language"
          },
          "detected_language_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Language Confidence"
          },
          "emotion_label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmotionLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "emotion_overridden": {
            "title": "Emotion Overridden",
            "type": "boolean"
          },
          "emotion_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotion Score"
          },
          "end_ms": {
            "title": "End Ms",
            "type": "integer"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "source_audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Audio Url"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Text"
          },
          "speaker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Id"
          },
          "start_ms": {
            "title": "Start Ms",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/SegmentStatus"
          },
          "sync_offset_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Offset Pct"
          },
          "translated_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translated Text"
          },
          "tts_audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tts Audio Url"
          },
          "tts_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tts Duration Ms"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "speaker_id",
          "index",
          "start_ms",
          "end_ms",
          "source_text",
          "detected_language",
          "detected_language_confidence",
          "translated_text",
          "emotion_label",
          "emotion_score",
          "emotion_overridden",
          "source_audio_url",
          "tts_audio_url",
          "tts_duration_ms",
          "sync_offset_pct",
          "status",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "SegmentRead",
        "type": "object"
      },
      "SegmentStatus": {
        "enum": [
          "pending",
          "transcribed",
          "emotion_detected",
          "translated",
          "synthesized",
          "muxed",
          "failed"
        ],
        "title": "SegmentStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend for Sur, an emotion-aware AI dubbing pipeline. REST under /api for project/segment CRUD and control; a WebSocket at /ws/projects/{id} for live pipeline progress. OpenAPI schema here is the contract the frontend generates its TypeScript types from (see frontend build prompt, PRD section 08).",
    "title": "Sur API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/segments/{segment_id}": {
      "patch": {
        "operationId": "patch_segment_api_segments__segment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "segment_id",
            "required": true,
            "schema": {
              "title": "Segment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Segment",
        "tags": [
          "segments"
        ]
      }
    },
    "/api/segments/{segment_id}/regenerate": {
      "post": {
        "description": "Queues a re-render of this segment only, then a re-mux of the export.\n\nThe project is marked processing (current_stage \"regenerate\") until the\nre-mux finishes, so the UI can show that work is in flight, the dashboard\ndownload is not the stale video, and liveness applies to it like any run.\nProgress and failures arrive on the project's WebSocket channel.",
        "operationId": "regenerate_segment_api_segments__segment_id__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "segment_id",
            "required": true,
            "schema": {
              "title": "Segment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Segment",
        "tags": [
          "segments"
        ]
      }
    },
    "/api/storage/files/{key}": {
      "get": {
        "operationId": "get_local_file_api_storage_files__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Local File",
        "tags": [
          "storage"
        ]
      },
      "head": {
        "operationId": "head_local_file_api_storage_files__key__head",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Head Local File",
        "tags": [
          "storage"
        ]
      }
    },
    "/api/storage/upload/{key}": {
      "put": {
        "operationId": "upload_local_file_api_storage_upload__key__put",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Local File",
        "tags": [
          "storage"
        ]
      }
    },
...
  }
}